load dynamic buttons from scripts SC support for dynamic buttion editing
